server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "sqlalchemy[asyncio]>=2.0.42",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "neo4j>=5.20.0",
    "pydantic>=2.5.0",
]
